+++ packages/web3-core-method/types/index.d.ts
@@ -32,3 +32,5 @@ export interface Method {
     abiCoder?: any;
     handleRevert?: boolean;
 }
+
+export default Method