+++ CHANGELOG.md
@@ -646,3 +646,4 @@ Released with 1.0.0-beta.37 code base.
 
   - Fixed skipped ws-ganache tests (#5759)
   - Fixed "provider started to reconnect error" in web3-provider-ws (#5820)
+  - Fixed types for packages which have default exports but not declared default export in .d.ts